move dag method out
